refactor: context implemented using usecontext
