+++ quiz/src/App.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useReducer, useState } from "react";
 import "./App.css";
 import { quizOne } from "./Data/getQuiz";
+import { useReduce } from "./Reducer-context";
 
 type Headerprops = {
   username: string;
@@ -23,54 +24,11 @@ type quizstate = {
   status: statustype;
   currentQsnNo: number;
 };
-const initialstate: quizstate = {
-  score: 0,
-  status: "starting",
-  currentQsnNo: 1
-};
 
-type actiontype =
-  | { type: "RESET" }
-  | { type: "INCREMENT_SCORE"; payload: { score: number } }
-  | { type: "DECREMENT_SCORE"; payload: { score: number } }
-  | { type: "SKIP" };
-
-function quizreducer(state: quizstate, action: actiontype): quizstate {
-  switch (action.type) {
-    case "RESET":
-      return { ...state, score: 0, status: "starting", currentQsnNo: 1 };
-
-    case "INCREMENT_SCORE":
-      return {
-        ...state,
-        score: state.score + action.payload.score,
-        currentQsnNo: state.currentQsnNo + 1
-      };
-
-    case "DECREMENT_SCORE":
-      return {
-        ...state,
-        score: state.score - action.payload.score,
-        currentQsnNo: state.currentQsnNo + 1
-      };
-    case "SKIP":
-      console.log({
-        ...state,
-        score: state.score,
-        currentQsnNo: state.currentQsnNo + 1
-      });
-      return {
-        ...state,
-        score: state.score,
-        currentQsnNo: state.currentQsnNo + 1
-      };
-    default:
-      return state;
-  }
-}
 console.log("lpop");
 
-function Qsnblock({ state, dispatch }: { state: quizstate; dispatch: any }) {
+function Qsnblock() {
+  let { state, dispatch } = useReduce();
   const [time, settime] = useState(15);
   console.log("qsnblock");
   useEffect(() => {
@@ -126,9 +84,8 @@ function Qsnblock({ state, dispatch }: { state: quizstate; dispatch: any }) {
 }
 export default function App() {
   console.log("render");
-  let [state, dispatch] = useReducer(quizreducer, initialstate);
-
-  //console.log("render");
+  let { state, dispatch } = useReduce();
+  console.log(useReduce());
   return (
     <div className="App">
       <Header username={"jdk"} score={state.score} />
@@ -139,7 +96,7 @@ export default function App() {
       <br />
 
       {state.currentQsnNo <= quizOne.questions.length ? (
-        <Qsnblock state={state} dispatch={dispatch} />
+        <Qsnblock />
       ) : (
         "thank you"
       )}
