--- quiz/src/Reducer-context.ts
@@ -1,11 +1,0 @@
-import { createContext, useContext } from "react";
-
-import { useState } from "react";
-type statustype = "starting" | "finished" | "reset";
-
-type quizstate = {
-  score: number;
-  status: statustype;
-  currentQsnNo: number;
-};
-export const Reducercontext = createContext({} as quizstate);
